requirements.txt
flask==3.0.3
flask-cors==4.0.1
yt-dlp==2024.8.6
google-generativeai==0.7.2
openai-whisper==20231117
torch==2.3.1
gunicorn==22.0.0
